# backend/src/main/java/com/habitbuilder/NutritionTracker/modules/auth/dto/ProfileResponse.java
package com.habitbuilder.NutritionTracker.modules.auth.dto;

import lombok.AllArgsConstructor;
import lombok.Data;

@Data
@AllArgsConstructor
public class ProfileResponse {
    private Long id;
    private String name;
    private String email;
    private String age;
    private String gender;
    private String timezone;
}